fix(VulkanObjectShader): fix destructor
Check if we are destroying a moved-from object by looking if VulkanPipeline has
a valid handle. This is scotch, but when the objects will be more stable I'll get
around to implement move-semantics properly(TM). For now this works good enough since
most of the heavy lifing is done by vulkan_raii.hpp.