feat(VulkanRenderer): add Frame struct
A Frame groups duplicated resources to support in-flight frame drawing.
Mainly it reduces clutter around the class, but also enable a nice
astraction: beginFrame() returns a FrameHandle (basically its index in the
array of frames) and every operation that needs to interact with it will
take said handle; endFrame() takes back this handle and finishes the
frame.