refactor(Application): inline make/destroyWindow
Since Application::run() is blocks and creates/destroys the window,
there is no way that for some reason it goes out of scope with the
window still up. Thus the destroy function is useless and can be inlined
(the destructor can simply be defaulted, but needs to be declared in the
header to avoid unique_ptr shenanigans). The window-creation function
can also be inlined for symmetry.

Moreover, inlining everything into run() fixes the fact that half of the
initialization/destruction was in a function while the other half was in
another.