feat(Entity): add Entity class
The Entity class is basically just a glorified Component hashmap. It
always has a Transform (which is stored separately mainly for
convenience) and new components can be emplaced or inserted by moving
an alwady allocated unique_ptr. Access is done through iterators for a
list of uniformily typed Components (key is the componentId).