refactor(seng): use GlobalUniform and move from Scene to Renderer
Move the GUBO layout handling from Scene to Renderer, together with
the layout that will be used for samplers.

While we are doing the big refactoring, also reinforce the descriptor
cache with some extra info, to allow to allocate multiple descriptors
with the same layout.

With the introduction of GlobalUniform, give a facelift to ObjectShader
with a lot of TODOs/FIXMEs in preparation for the full shader system.